move aggregation client and types into extensions
