+++ stac_fastapi/extensions/stac_fastapi/extensions/core/aggregation/aggregation.py
@@ -7,9 +7,9 @@ from fastapi import APIRouter, FastAPI
 
 from stac_fastapi.api.models import CollectionUri, EmptyRequest
 from stac_fastapi.api.routes import create_async_endpoint
-from stac_fastapi.types.core import AsyncBaseAggregationClient, BaseAggregationClient
 from stac_fastapi.types.extension import ApiExtension
 
+from .client import AsyncBaseAggregationClient, BaseAggregationClient
 from .request import AggregationExtensionGetRequest, AggregationExtensionPostRequest
