+++ stac_fastapi/types/stac_fastapi/types/core.py
@@ -16,7 +16,6 @@ from stac_pydantic.shared import BBox, MimeTypes
 from starlette.responses import Response
 
 from stac_fastapi.types import stac
-from stac_fastapi.types.aggregation import Aggregation, AggregationCollection
 from stac_fastapi.types.config import ApiSettings
 from stac_fastapi.types.conformance import BASE_CONFORMANCE_CLASSES
 from stac_fastapi.types.extension import ApiExtension
@@ -799,129 +798,4 @@ def __getattr__(name: str) -> Any:
         clients = importlib.import_module("stac_fastapi.extensions.core.filter.client")
         return getattr(clients, name)
 
-
     raise AttributeError(f"module {__name__} has no attribute {name}")
-
-        This base implementation returns a blank queryable schema. This is not allowed
-        under OGC CQL but it is allowed by the STAC API Filter Extension
-        https://github.com/stac-api-extensions/filter#queryables
-        """
-        return {
-            "$schema": "https://json-schema.org/draft/2019-09/schema",
-            "$id": "https://example.org/queryables",
-            "type": "object",
-            "title": "Queryables for Example STAC API",
-            "description": "Queryable names for the example STAC API Item Search filter.",
-            "properties": {},
-        }
-
-
-@attr.s
-class BaseAggregationClient(abc.ABC):
-    """Defines a pattern for implementing the STAC aggregation extension."""
-
-    def get_aggregations(
-        self, collection_id: Optional[str] = None, **kwargs
-    ) -> AggregationCollection:
-        """Get the aggregations available for the given collection_id.
-
-        If collection_id is None, returns the available aggregations over all
-        collections.
-        """
-        return AggregationCollection(
-            type="AggregationCollection",
-            aggregations=[Aggregation(name="total_count", data_type="integer")],
-            links=[
-                {
-                    "rel": "root",
-                    "type": "application/json",
-                    "href": "https://example.org/",
-                },
-                {
-                    "rel": "self",
-                    "type": "application/json",
-                    "href": "https://example.org/aggregations",
-                },
-            ],
-        )
-
-    def aggregate(
-        self, collection_id: Optional[str] = None, **kwargs
-    ) -> AggregationCollection:
-        """Return the aggregation buckets for a given search result"""
-        return AggregationCollection(
-            type="AggregationCollection",
-            aggregations=[],
-            links=[
-                {
-                    "rel": "root",
-                    "type": "application/json",
-                    "href": "https://example.org/",
-                },
-                {
-                    "rel": "self",
-                    "type": "application/json",
-                    "href": "https://example.org/aggregations",
-                },
-            ],
-        )
-
-
-@attr.s
-class AsyncBaseAggregationClient(abc.ABC):
-    """Defines an async pattern for implementing the STAC aggregation extension."""
-
-    async def get_aggregations(
-        self, collection_id: Optional[str] = None, **kwargs
-    ) -> AggregationCollection:
-        """Get the aggregations available for the given collection_id.
-
-        If collection_id is None, returns the available aggregations over all
-        collections.
-        """
-        return AggregationCollection(
-            type="AggregationCollection",
-            aggregations=[Aggregation(name="total_count", data_type="integer")],
-            links=[
-                {
-                    "rel": "root",
-                    "type": "application/json",
-                    "href": "https://example.org/",
-                },
-                {
-                    "rel": "self",
-                    "type": "application/json",
-                    "href": "https://example.org/aggregations",
-                },
-            ],
-        )
-
-    async def aggregate(
-        self,
-        collection_id: Optional[str] = None,
-        aggregations: Optional[Union[str, List[str]]] = None,
-        collections: Optional[List[str]] = None,
-        ids: Optional[List[str]] = None,
-        bbox: Optional[BBox] = None,
-        intersects: Optional[Geometry] = None,
-        datetime: Optional[DateTimeType] = None,
-        limit: Optional[int] = 10,
-        **kwargs,
-    ) -> AggregationCollection:
-        """Return the aggregation buckets for a given search result"""
-        return AggregationCollection(
-            type="AggregationCollection",
-            aggregations=[],
-            links=[
-                {
-                    "rel": "root",
-                    "type": "application/json",
-                    "href": "https://example.org/",
-                },
-                {
-                    "rel": "self",
-                    "type": "application/json",
-                    "href": "https://example.org/aggregations",
-                },
-            ],
-        )
